--- packages/core/src/extensions/Comments/threadstore/LiveBlocksThreadStore.ts.bak
@@ -1,8 +1,0 @@
-export class LiveblocksThreadStore {
-  constructor(private readonly editor: BlockNoteEditor<any, any, any>) {}
-
-  public async createThread() {
-    const x = useCreateThread();
-    return x;
-  }
-}
