+++ packages/core/src/extensions/Comments/threadstore/yjs/YjsThreadStore.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import * as Y from "yjs";
-import { CommentBody } from "../types.js";
-import { DefaultThreadStoreAuth } from "./DefaultThreadStoreAuth.js";
+import { CommentBody } from "../../types.js";
+import { DefaultThreadStoreAuth } from "../DefaultThreadStoreAuth.js";
 import { YjsThreadStore } from "./YjsThreadStore.js";
 
 // Mock UUID to generate sequential IDs
